contrastive/visualization/gif_umap/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
umap-learn>=0.5.2
plotly>=5.3.0
jinja2>=3.0.0 